refactor getCorrectAnswer function
